requirements.txt
pandas
psycopg2-binary
python-dateutil
matplotlib
seaborn